Edit quote model

+++ database/migrations/2024_03_13_165842_create_quotes_table.php
@@ -22,9 +22,9 @@ return new class extends Migration
             $table->foreignIdFor(User::class)->constrained()->cascadeOnDelete();
             $table->foreignIdFor(Vertical::class)->constrained()->cascadeOnDelete();
             $table->string('currency')->default('KES');
-            $table->bigInteger('subtotal');
-            $table->bigInteger('taxes');
-            $table->bigInteger('total');
+            $table->integer('subtotal');
+            $table->integer('taxes');
+            $table->integer('total');
             $table->enum('series', QuoteSeries::values())->default(QuoteSeries::IN2QUT->name);
             $table->integer('serial_number')->nullable();
             $table->string('serial')->nullable();
